Map: move geo_points getting to geo points layer, remove map#index action - using geo_points#index instead, update tests refs: #ZN-44
