+++ app/controllers/api/v1/map_controller.rb
@@ -3,12 +3,6 @@
 module Api
   module V1
     class MapController < Api::V1::ApplicationController
-      def index
-        # render here geo_points/users/all related to maps info
-        geo_points = GeoPoint.all.map(&:json)
-
-        render json: { geoPoints: geo_points }.to_json
-      end
     end
   end
 end
