+++ app/frontend/javascript/components/map/components/map_layers/overlays/geo_points_layer.jsx
@@ -1,7 +1,35 @@
+import { connect } from "react-redux";
+import { setGeoPoints } from "../../../../../store/actions/geo_points";
+import { setMainMapBlock } from "../../../../../store/actions/main_map";
+import fetchLink from "../../../../../helpers/fetch_link";
 import { FeatureGroup } from "react-leaflet";
 import GeoPointMarker from "../../geo_point_marker";
 
 class GeoPointsLayer extends React.Component {
+  async toggleMapBlock (state, blockMessage) {
+    this.props.setMainMapBlock({ state, blockMessage });
+  }
+
+  componentDidMount () {
+    this.toggleMapBlock(true, I18n.t("common.screen_block.loading_geo_points"));
+
+    fetchLink({
+      link: "/api/v1/geo_points",
+      onSuccess: (response) => {
+        this.props.setGeoPoints(
+          response.geoPoints.map(({ data: { attributes: { id, user_id: userId, longitude, latitude, rad_value: radValue, comment } } }) => (
+            { id, userId, longitude, latitude, radValue, comment }
+          ))
+        );
+      },
+      onFailure: (error) => {
+        // TODO: add parsing of internal server errors
+        throw new Error(error);
+      },
+      onComplete: () => this.toggleMapBlock(false)
+    });
+  }
+
   renderGeoPoints () {
     const { geoPoints, currentUserId } = this.props;
 
@@ -29,9 +57,11 @@ class GeoPointsLayer extends React.Component {
   }
 }
 
-GeoPointsLayer.propTypes = {
-  geoPoints: PropTypes.array.isRequired,
-  currentUserId: PropTypes.number
-}
+const mapStateToProps = ({ geoPoints, currentUser: { id: currentUserId } }) => ({ geoPoints, currentUserId });
 
-export default GeoPointsLayer;
+const mapDispatchToProps = dispatch => ({
+  setGeoPoints: (geoPoints) => dispatch(setGeoPoints(geoPoints)),
+  setMainMapBlock: (block) => dispatch(setMainMapBlock(block))
+});
+
+export default connect(mapStateToProps, mapDispatchToProps)(GeoPointsLayer);
