+++ config/routes.rb
@@ -12,7 +12,6 @@ Rails.application.routes.draw do
     namespace :v1 do
       resources :current_user, only: :index
       resources :geo_points, except: %i[new edit]
-      resources :map, only: :index
     end
   end
